requirements.txt
python-dotenv
pyaudio
pygame
pydub
SpeechRecognition
assemblyai
openai
